server js changes

+++ app/Http/Controllers/Api/AdController.php
@@ -1136,8 +1136,8 @@ class AdController extends Controller
             'status'=>config('global.NOT_FOUND_STATUS')
         ],config('global.NOT_FOUND_STATUS'));
 
-        return gettype($request->social_media);
-        if(count($request->social_media) > 0)
+        
+        if(gettype($request->social_media) == 'array'&&count($request->social_media) > 0)
         {
             DB::table('social_media_id')->where('ad_id',$data->id)->delete();
